me when types

+++ src/generic_back/arrangement.rs
@@ -50,9 +50,7 @@ impl Arrangement {
         .unwrap();
 
         (0..self.len().in_interleaved_samples(meter)).for_each(|i| {
-            writer
-                .write_sample(self.get_at_global_time(i as u32))
-                .unwrap();
+            writer.write_sample(self.get_at_global_time(i)).unwrap();
         });
 
         self.meter.playing.store(false, SeqCst);
